+++ cli/src/main.cpp
@@ -20,7 +20,7 @@ std::unique_ptr<subcalls::subcall> create_subcall(const std::string& subcall) {
     }
 }
 
-void print_general_help(cxxopts::Options& options) {
+void print_general_help(const cxxopts::Options& options) {
     std::cout << options.help() << "\n";
     std::cout << "Available subcommands: \n";
     std::cout << "\tidx:\t\tIndex an Interval File\n";